+++ src/util/number.rs
@@ -5,11 +5,11 @@ use byteorder::{BigEndian, ReadBytesExt};
 
 pub struct NumberBytes {
     tail: Vec<u8>,
-    number_width: usize,
+    number_width: u8,
 }
 
 impl NumberBytes {
-    pub fn new(bytes: &Vec<u8>, number_width: usize) -> Self {
+    pub fn new(bytes: &Vec<u8>, number_width: u8) -> Self {
         NumberBytes {
             tail: bytes.clone(),
             number_width,
@@ -18,10 +18,10 @@ impl NumberBytes {
 
     pub fn read_number(&mut self) -> Option<u64> {
         let cur_tail = self.tail.clone();
-        if cur_tail.len() < self.number_width {
+        if cur_tail.len() < self.number_width as usize {
             return None;
         }
-        let (mut num, tail_bytes) = cur_tail.split_at(self.number_width);
+        let (mut num, tail_bytes) = cur_tail.split_at(self.number_width as usize);
         self.tail = Vec::from(tail_bytes);
         Some(num.read_u64::<BigEndian>().unwrap())
     }
